Card storage
Need to bring players deck into battle now

+++ Super_Serious_Space_Game/assets/Scripts/StaticScripts/PlayerInfoStatic.cs
@@ -29,7 +29,7 @@ public class PlayerInfoStatic : MonoBehaviour {
     public static Sprite HomePlanetSprite;
     public static Sprite ShipImage;
 
-    public static float CardID;
+    public static int CardID;
     public static List<GameObject> PlayerUnitCards;
     public static List<GameObject> PlayerDeck;
 }
